sea_orm

Derive Macro EnumIter

#[derive(EnumIter)]
{
    // Attributes available to this derive:
    #[strum]
}

Creates a new type that iterates of the variants of an enum.

Iterate over the variants of an Enum. Any additional data on your variants will be set to Default::default(). The macro implements strum::IntoEnumIterator on your enum and creates a new type called YourEnumIter that is the iterator object. You cannot derive EnumIter on any type with a lifetime bound (<'a>) because the iterator would surely create unbounded lifetimes.