sea_orm::entity::prelude

Struct TimeTime

#[repr(C)]
pub struct TimeTime { /* private fields */ }

The clock time within a given date. Nanosecond precision.

All minutes are assumed to have exactly 60 seconds; no attempt is made to handle leap seconds (either positive or negative).

When comparing two Times, they are assumed to be in the same calendar date.

Implementations

impl Time

pub const MIDNIGHT: Time

A Time that is exactly midnight. This is the smallest possible value for a Time.

assert_eq!(Time::MIDNIGHT, time!(0:00));

pub const MAX: Time

A Time that is one nanosecond before midnight. This is the largest possible value for a Time.

assert_eq!(Time::MAX, time!(23:59:59.999_999_999));

pub const fn from_hms( hour: u8, minute: u8, second: u8, ) -> Result<Time, ComponentRange>

Attempt to create a Time from the hour, minute, and second.

assert!(Time::from_hms(1, 2, 3).is_ok());

assert!(Time::from_hms(24, 0, 0).is_err()); // 24 isn't a valid hour.
assert!(Time::from_hms(0, 60, 0).is_err()); // 60 isn't a valid minute.
assert!(Time::from_hms(0, 0, 60).is_err()); // 60 isn't a valid second.

pub const fn from_hms_milli( hour: u8, minute: u8, second: u8, millisecond: u16, ) -> Result<Time, ComponentRange>

Attempt to create a Time from the hour, minute, second, and millisecond.

assert!(Time::from_hms_milli(1, 2, 3, 4).is_ok());

assert!(Time::from_hms_milli(24, 0, 0, 0).is_err()); // 24 isn't a valid hour.
assert!(Time::from_hms_milli(0, 60, 0, 0).is_err()); // 60 isn't a valid minute.
assert!(Time::from_hms_milli(0, 0, 60, 0).is_err()); // 60 isn't a valid second.
assert!(Time::from_hms_milli(0, 0, 0, 1_000).is_err()); // 1_000 isn't a valid millisecond.

pub const fn from_hms_micro( hour: u8, minute: u8, second: u8, microsecond: u32, ) -> Result<Time, ComponentRange>

Attempt to create a Time from the hour, minute, second, and microsecond.

assert!(Time::from_hms_micro(1, 2, 3, 4).is_ok());

assert!(Time::from_hms_micro(24, 0, 0, 0).is_err()); // 24 isn't a valid hour.
assert!(Time::from_hms_micro(0, 60, 0, 0).is_err()); // 60 isn't a valid minute.
assert!(Time::from_hms_micro(0, 0, 60, 0).is_err()); // 60 isn't a valid second.
assert!(Time::from_hms_micro(0, 0, 0, 1_000_000).is_err()); // 1_000_000 isn't a valid microsecond.

pub const fn from_hms_nano( hour: u8, minute: u8, second: u8, nanosecond: u32, ) -> Result<Time, ComponentRange>

Attempt to create a Time from the hour, minute, second, and nanosecond.

assert!(Time::from_hms_nano(1, 2, 3, 4).is_ok());

assert!(Time::from_hms_nano(24, 0, 0, 0).is_err()); // 24 isn't a valid hour.
assert!(Time::from_hms_nano(0, 60, 0, 0).is_err()); // 60 isn't a valid minute.
assert!(Time::from_hms_nano(0, 0, 60, 0).is_err()); // 60 isn't a valid second.
assert!(Time::from_hms_nano(0, 0, 0, 1_000_000_000).is_err()); // 1_000_000_000 isn't a valid nanosecond.

pub const fn as_hms(self) -> (u8, u8, u8)

Get the clock hour, minute, and second.

assert_eq!(time!(0:00:00).as_hms(), (0, 0, 0));
assert_eq!(time!(23:59:59).as_hms(), (23, 59, 59));

pub const fn as_hms_milli(self) -> (u8, u8, u8, u16)

Get the clock hour, minute, second, and millisecond.

assert_eq!(time!(0:00:00).as_hms_milli(), (0, 0, 0, 0));
assert_eq!(time!(23:59:59.999).as_hms_milli(), (23, 59, 59, 999));

pub const fn as_hms_micro(self) -> (u8, u8, u8, u32)

Get the clock hour, minute, second, and microsecond.

assert_eq!(time!(0:00:00).as_hms_micro(), (0, 0, 0, 0));
assert_eq!(
    time!(23:59:59.999_999).as_hms_micro(),
    (23, 59, 59, 999_999)
);

pub const fn as_hms_nano(self) -> (u8, u8, u8, u32)

Get the clock hour, minute, second, and nanosecond.

assert_eq!(time!(0:00:00).as_hms_nano(), (0, 0, 0, 0));
assert_eq!(
    time!(23:59:59.999_999_999).as_hms_nano(),
    (23, 59, 59, 999_999_999)
);

pub const fn hour(self) -> u8

Get the clock hour.

The returned value will always be in the range 0..24.

assert_eq!(time!(0:00:00).hour(), 0);
assert_eq!(time!(23:59:59).hour(), 23);

pub const fn minute(self) -> u8

Get the minute within the hour.

The returned value will always be in the range 0..60.

assert_eq!(time!(0:00:00).minute(), 0);
assert_eq!(time!(23:59:59).minute(), 59);

pub const fn second(self) -> u8

Get the second within the minute.

The returned value will always be in the range 0..60.

assert_eq!(time!(0:00:00).second(), 0);
assert_eq!(time!(23:59:59).second(), 59);

pub const fn millisecond(self) -> u16

Get the milliseconds within the second.

The returned value will always be in the range 0..1_000.

assert_eq!(time!(0:00).millisecond(), 0);
assert_eq!(time!(23:59:59.999).millisecond(), 999);

pub const fn microsecond(self) -> u32

Get the microseconds within the second.

The returned value will always be in the range 0..1_000_000.

assert_eq!(time!(0:00).microsecond(), 0);
assert_eq!(time!(23:59:59.999_999).microsecond(), 999_999);

pub const fn nanosecond(self) -> u32

Get the nanoseconds within the second.

The returned value will always be in the range 0..1_000_000_000.

assert_eq!(time!(0:00).nanosecond(), 0);
assert_eq!(time!(23:59:59.999_999_999).nanosecond(), 999_999_999);

pub const fn duration_until(self, other: Time) -> Duration

Determine the Duration that, if added to self, would result in the parameter.

assert_eq!(time!(18:00).duration_until(Time::MIDNIGHT), 6.hours());
assert_eq!(time!(23:00).duration_until(time!(1:00)), 2.hours());

pub const fn duration_since(self, other: Time) -> Duration

Determine the Duration that, if added to the parameter, would result in self.

assert_eq!(Time::MIDNIGHT.duration_since(time!(18:00)), 6.hours());
assert_eq!(time!(1:00).duration_since(time!(23:00)), 2.hours());

pub const fn replace_hour(self, hour: u8) -> Result<Time, ComponentRange>

Replace the clock hour.

assert_eq!(
    time!(01:02:03.004_005_006).replace_hour(7),
    Ok(time!(07:02:03.004_005_006))
);
assert!(time!(01:02:03.004_005_006).replace_hour(24).is_err()); // 24 isn't a valid hour

pub const fn replace_minute(self, minute: u8) -> Result<Time, ComponentRange>

Replace the minutes within the hour.

assert_eq!(
    time!(01:02:03.004_005_006).replace_minute(7),
    Ok(time!(01:07:03.004_005_006))
);
assert!(time!(01:02:03.004_005_006).replace_minute(60).is_err()); // 60 isn't a valid minute

pub const fn replace_second(self, second: u8) -> Result<Time, ComponentRange>

Replace the seconds within the minute.

assert_eq!(
    time!(01:02:03.004_005_006).replace_second(7),
    Ok(time!(01:02:07.004_005_006))
);
assert!(time!(01:02:03.004_005_006).replace_second(60).is_err()); // 60 isn't a valid second

pub const fn replace_millisecond( self, millisecond: u16, ) -> Result<Time, ComponentRange>

Replace the milliseconds within the second.

assert_eq!(
    time!(01:02:03.004_005_006).replace_millisecond(7),
    Ok(time!(01:02:03.007))
);
assert!(time!(01:02:03.004_005_006)
    .replace_millisecond(1_000)
    .is_err()); // 1_000 isn't a valid millisecond

pub const fn replace_microsecond( self, microsecond: u32, ) -> Result<Time, ComponentRange>

Replace the microseconds within the second.

assert_eq!(
    time!(01:02:03.004_005_006).replace_microsecond(7_008),
    Ok(time!(01:02:03.007_008))
);
assert!(time!(01:02:03.004_005_006)
    .replace_microsecond(1_000_000)
    .is_err()); // 1_000_000 isn't a valid microsecond

pub const fn replace_nanosecond( self, nanosecond: u32, ) -> Result<Time, ComponentRange>

Replace the nanoseconds within the second.

assert_eq!(
    time!(01:02:03.004_005_006).replace_nanosecond(7_008_009),
    Ok(time!(01:02:03.007_008_009))
);
assert!(time!(01:02:03.004_005_006)
    .replace_nanosecond(1_000_000_000)
    .is_err()); // 1_000_000_000 isn't a valid nanosecond

impl Time

pub fn format_into( self, output: &mut (impl Write + ?Sized), format: &(impl Formattable + ?Sized), ) -> Result<usize, Format>

Format the Time using the provided format description.

pub fn format( self, format: &(impl Formattable + ?Sized), ) -> Result<String, Format>

Format the Time using the provided format description.

let format = format_description::parse("[hour]:[minute]:[second]")?;
assert_eq!(time!(12:00).format(&format)?, "12:00:00");

impl Time

pub fn parse( input: &str, description: &(impl Parsable + ?Sized), ) -> Result<Time, Parse>

Parse a Time from the input using the provided format description.

let format = format_description!("[hour]:[minute]:[second]");
assert_eq!(Time::parse("12:00:00", &format)?, time!(12:00));