sea_orm

Module error

Error types for all database operations

Structs

ColumnFromStrErr

Error during impl FromStr for Entity::Column

SqlxMySqlError

An error returned from the MySQL database.

SqlxPostgresError

An error returned from the PostgreSQL database.

SqlxSqliteError

Enums

ConnAcquireErr

Connection Acquire error

DbErr

An error from unsuccessful database operations

RuntimeErr

Runtime error

SqlErr

An error from unsuccessful SQL query

SqlxError

Represents all the ways a method can fail within SQLx.

TryGetError

An error from trying to get a row from a Model