Module sea_orm::query

Holds types and methods to perform queries

Macros

debug_query

Helper to get a raw SQL string from an object that impl QueryTrait.

debug_query_stmt

Helper to get a Statement from an object that impl QueryTrait.

Structs

Condition

Represents the value of an Condition::any or Condition::all: a set of disjunctive or conjunctive conditions.

DebugQuery

This structure provides debug capabilities

Delete

Defines the structure for a delete operation

DeleteMany

Perform a delete operation on multiple models

DeleteOne

Perform a delete operation on a model

Insert

Performs INSERT operations on a ActiveModel

InsertResult

The result of an INSERT operation on an ActiveModel

Select

Defines a structure to perform select operations

SelectA

Implements the traits Iden and IdenStatic for a type

SelectB

Implements the traits Iden and IdenStatic for a type

SelectTwo

Defines a structure to perform a SELECT operation on two Models

SelectTwoMany

Defines a structure to perform a SELECT operation on many Models

Statement

Defines an SQL statement

Update

Defines a structure to perform UPDATE query operations on a ActiveModel

UpdateMany

Defines an UPDATE operation on multiple ActiveModels

UpdateOne

Defines an UPDATE operation on one ActiveModel

UpdateResult

The result of an update operation on an ActiveModel

Values

Enums

JoinType

Join types

JsonValue

Represents any valid JSON value.

Order

Ordering options

Value

Value variants

Traits

ConditionalStatement

ConnectionTrait

Creates constraints for any structure that can create a database connection and execute SQL statements

CursorTrait

A trait for any type that can be turn into a cursor

IntoSimpleExpr

Performs a conversion to SimpleExpr

OrderedStatement

PaginatorTrait

A Trait for any type that can paginate results

QueryFilter

Perform a FILTER opertation on a statement

QueryOrder

Performs ORDER BY operations

QuerySelect

Constraints for any type that needs to perform select statements on a Model

QueryTrait

A Trait for any type performing queries on a Model or ActiveModel

StreamTrait

Stream query results

TransactionTrait

Spawn database transaction

Type Definitions

DynIden