Struct sea_orm::ConnectOptions

pub struct ConnectOptions { /* private fields */ }

Defines the configuration options of a database

Implementations

impl ConnectOptions

pub fn new(url: String) -> Self

Create new ConnectOptions for a Database by passing in a URI string

pub fn pool_options<DB>(self) -> PoolOptions<DB>where
DB: Database,

Convert ConnectOptions into sqlx::pool::PoolOptions

pub fn get_url(&self) -> &str

Get the database URL of the pool

pub fn max_connections(&mut self, value: u32) -> &mut Self

Set the maximum number of connections of the pool

pub fn get_max_connections(&self) -> Option<u32>

Get the maximum number of connections of the pool, if set

pub fn min_connections(&mut self, value: u32) -> &mut Self

Set the minimum number of connections of the pool

pub fn get_min_connections(&self) -> Option<u32>

Get the minimum number of connections of the pool, if set

pub fn connect_timeout(&mut self, value: Duration) -> &mut Self

Set the timeout duration when acquiring a connection

pub fn get_connect_timeout(&self) -> Option<Duration>

Get the timeout duration when acquiring a connection, if set

pub fn idle_timeout(&mut self, value: Duration) -> &mut Self

Set the idle duration before closing a connection

pub fn get_idle_timeout(&self) -> Option<Duration>

Get the idle duration before closing a connection, if set

pub fn acquire_timeout(&mut self, value: Duration) -> &mut Self

Set the maximum amount of time to spend waiting for acquiring a connection

pub fn get_acquire_timeout(&self) -> Option<Duration>

Get the maximum amount of time to spend waiting for acquiring a connection

pub fn max_lifetime(&mut self, lifetime: Duration) -> &mut Self

Set the maximum lifetime of individual connections

pub fn get_max_lifetime(&self) -> Option<Duration>

Get the maximum lifetime of individual connections, if set

pub fn sqlx_logging(&mut self, value: bool) -> &mut Self

Enable SQLx statement logging (default true)

pub fn get_sqlx_logging(&self) -> bool

Get whether SQLx statement logging is enabled

pub fn sqlx_logging_level(&mut self, level: LevelFilter) -> &mut Self

Set SQLx statement logging level (default INFO) (ignored if sqlx_logging is false)

pub fn get_sqlx_logging_level(&self) -> LevelFilter

Get the level of SQLx statement logging

pub fn sqlcipher_key<T>(&mut self, value: T) -> &mut Selfwhere
T: Into<Cow<'static, str>>,

set key for sqlcipher

pub fn set_schema_search_path(&mut self, schema_search_path: String) -> &mut Self

Set schema search path (PostgreSQL only)