Trait sea_orm::SelectorTrait

pub trait SelectorTrait {
    type Item: Sized;

    fn from_raw_query_result(res: QueryResult) -> Result<Self::Item, DbErr>;
}

A Trait for any type that can perform SELECT queries

Required Associated Types

type Item: Sized

Required Methods

fn from_raw_query_result(res: QueryResult) -> Result<Self::Item, DbErr>

The method to perform a query on a Model