sea_orm

Module entity

Module for the Entity type and operations

Re-exports

pub use compound::EntityLoaderTrait;

pub use ActiveValue::NotSet;

pub use ActiveValue::Set;

pub use ActiveValue::Unchanged;

Modules

compound

prelude

Re-export common types from the entity

Structs

BoolColumn

BorrowedIdentityIter

Iterator for Identity

BytesColumn

ColumnDef

Defines a Column for an Entity

DateLikeColumn

Supports both chrono and time Date

DateTimeLikeColumn

Supports both chrono and time DateTime

GenericArrayColumn

IpNetworkColumn

JsonColumn

NumericArrayColumn

NumericColumn

A column of numeric type, including integer, float and decimal

NumericColumnNullable

A column of numeric type, including integer, float and decimal that is also nullable

OwnedIdentityIter

Iterator for Identity

RelationBuilder

Defines a helper to build a relation

RelationDef

Defines a relationship

SeaographyColumnAttr

Seaography specific attributes

StringColumn

StringColumnNullable

TextUuidColumn

TimeLikeColumn

Supports both chrono and time Time

UuidColumn

Enums

ActiveModelAction

Action to perform on ActiveModel

ActiveValue

The state of a field in an [ActiveModel][ActiveModelTrait].

ColumnType

All column types

HasManyModel

Container for 1-N or M-N related Models

HasOneModel

Container for belongs_to or has_one relation

Identity

List of column identifier

RelationType

Defines the type of relationship

Value

Value variants

Traits

ActiveEnum

A Rust representation of enum defined in database.

ActiveEnumValue

The Rust Value backing ActiveEnums

ActiveModelBehavior

A Trait for overriding the ActiveModel behavior

ActiveModelTrait

ActiveModel is a type for constructing INSERT and UPDATE statements for a particular table.

ColumnTrait

API for working with a Column. Mostly a wrapper of the identically named methods in sea_query::Expr

ColumnTypeTrait

SeaORM’s utility methods that act on ColumnType

EntityName

A Trait for mapping an Entity to a database table

EntityTrait

An abstract base class for defining Entities.

FromQueryResult

A Trait for implementing a QueryResult

IdenStatic

Ensure the identifier for an Entity can be converted to a static str

IdentityOf

Check the Identity of an Entity

IntoActiveModel

A Trait for any type that can be converted into an ActiveModel

IntoActiveValue

Any type that can be converted into an ActiveValue

IntoIdentity

Performs a conversion into an Identity

IntoOption

Iterable

This trait designates that an Enum can be iterated over. It can be auto generated using the EnumIter derive macro.

Linked

A Trait for links between Entities

ModelTrait

The interface for Model, implemented by data structs

PartialModelTrait

A trait for a part of Model

PrimaryKeyArity

How many columns this Primary Key comprises

PrimaryKeyToColumn

Trait to map a Primary Key to a column

PrimaryKeyTrait

A Trait for to be used to define a Primary Key.

Related

A trait to relate two Entities for them to be joined in queries

RelatedSelfVia

A trait to relate an Entity to itself through a junction table

RelationTrait

Defines the relations of an Entity

TryIntoModel

A Trait for any type that can be converted into an Model

Functions

UnsetDeprecated

Defines an not set operation on an ActiveValue

Type Aliases

ForeignKeyAction

Action to perform on a foreign key whenever there are changes to an ActiveModel

LinkDef

Same as RelationDef