sea_orm::entity
Module prelude

Re-export common types from the entity

Re-exports

pub use crate::ActiveEnum;

pub use crate::ActiveModelBehavior;

pub use crate::ActiveModelTrait;

pub use crate::ColumnDef;

pub use crate::ColumnTrait;

pub use crate::ColumnTypeTrait;

pub use crate::ConnectionTrait;

pub use crate::CursorTrait;

pub use crate::DatabaseConnection;

pub use crate::DbConn;

pub use crate::EntityName;

pub use crate::EntityTrait;

pub use crate::ForeignKeyAction;

pub use crate::IdenStatic;

pub use crate::Linked;

pub use crate::LoaderTrait;

pub use crate::ModelTrait;

pub use crate::PaginatorTrait;

pub use crate::PrimaryKeyArity;

pub use crate::PrimaryKeyToColumn;

pub use crate::PrimaryKeyTrait;

pub use crate::QueryFilter;

pub use crate::QueryResult;

pub use crate::Related;

pub use crate::RelationDef;

pub use crate::RelationTrait;

pub use crate::Select;

pub use async_trait;

pub use crate::error::*;

Structs

BigDecimal

A big decimal type.

ChronoDate

ISO 8601 calendar date without timezone. Allows for every proleptic Gregorian date from Jan 1, 262145 BCE to Dec 31, 262143 CE. Also supports the conversion from ISO 8601 ordinal and week date.

ChronoDateTime

ISO 8601 combined date and time without timezone.

ChronoTime

ISO 8601 time without timezone. Allows for the nanosecond precision and optional leap second representation.

Date

ISO 8601 calendar date without timezone. Allows for every proleptic Gregorian date from Jan 1, 262145 BCE to Dec 31, 262143 CE. Also supports the conversion from ISO 8601 ordinal and week date.

DateTime

ISO 8601 combined date and time without timezone.

Decimal

Decimal represents a 128 bit representation of a fixed-precision decimal number. The finite set of values of type Decimal are of the form m / 10e, where m is an integer such that -296 < m < 296, and e is an integer between 0 and 28 inclusive.

Expr

Helper to build a SimpleExpr.

RcOrArc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

SeaRc

Time

ISO 8601 time without timezone. Allows for the nanosecond precision and optional leap second representation.

TimeDate

Date in the proleptic Gregorian calendar.

TimeDateTime

Combined date and time.

TimeDateTimeWithTimeZone

A PrimitiveDateTime with a UtcOffset.

TimeTime

The clock time within a given date. Nanosecond precision.

Uuid

A Universally Unique Identifier (UUID).

Enums

ColumnType

All column types

Json

Represents any valid JSON value.

StringLen

Length for var-char/binary; default to 255

Value

Value variants

Traits

Iden

Identifier

OverrideTableSchema

Override table name if database schema is set

Functions

override_table_schema_str

Override table name if database schema is set

Type Aliases

ChronoDateTimeLocal

Date time represented in local time

ChronoDateTimeUtc

Date time represented in UTC

ChronoDateTimeWithTimeZone

Date time with fixed offset

DateTimeLocal

Date time represented in local time

DateTimeUtc

Date time represented in UTC

DateTimeWithTimeZone

Date time with fixed offset

DynIden

Derive Macros

DeriveActiveEnum

A derive macro to implement sea_orm::ActiveEnum trait for enums.

DeriveActiveModel

The DeriveActiveModel derive macro will implement ActiveModelTrait for ActiveModel which provides setters and getters for all active values in the active model.

DeriveActiveModelBehavior

Models that a user can override

DeriveColumn

The DeriveColumn derive macro will implement [ColumnTrait] for Columns. It defines the identifier of each column by implementing Iden and IdenStatic. The EnumIter is also derived, allowing iteration over all enum variants.

DeriveCustomColumn

Derive a column if column names are not in snake-case

DeriveDisplay

DeriveEntity

Create an Entity

DeriveEntityModel

This derive macro is the ‘almighty’ macro which automatically generates Entity, Column, and PrimaryKey from a given Model.

DeriveIden

The DeriveIden derive macro will implement sea_orm::sea_query::Iden for simplify Iden implementation.

DeriveIntoActiveModel

Derive into an active model

DeriveModel

The DeriveModel derive macro will implement ModelTrait for Model, which provides setters and getters for all attributes in the mod It also implements FromQueryResult to convert a query result into the corresponding Model.

DerivePartialModel

The DerivePartialModel derive macro will implement sea_orm::PartialModelTrait for simplify partial model queries.

DerivePrimaryKey

The DerivePrimaryKey derive macro will implement [PrimaryKeyToColumn] for PrimaryKey which defines tedious mappings between primary keys and columns. The EnumIter is also derived, allowing iteration over all enum variants.

DeriveRelatedEntity

The DeriveRelatedEntity derive macro will implement seaography::RelationBuilder for RelatedEntity enumeration.

DeriveRelation

The DeriveRelation derive macro will implement RelationTrait for Relation.

DeriveValueType

Implements traits for types that wrap a database value type.

EnumIter

Creates a new type that iterates of the variants of an enum.