sea_orm::entity

Trait ModelTrait

pub trait ModelTrait:
    Clone
    + Send
    + Debug {
    type Entity: EntityTrait;

    // Required methods
    fn get(&self, c: <Self::Entity as EntityTrait>::Column) -> Value;
    fn get_value_type(c: <Self::Entity as EntityTrait>::Column) -> ArrayType;
    fn try_set(
        &mut self,
        c: <Self::Entity as EntityTrait>::Column,
        v: Value,
    ) -> Result<(), DbErr>;

    // Provided methods
    fn set(&mut self, c: <Self::Entity as EntityTrait>::Column, v: Value) { ... }
    fn find_related<R>(&self, _: R) -> Select<R>
       where R: EntityTrait,
             Self::Entity: Related<R> { ... }
    fn find_linked<L>(&self, l: L) -> Select<L::ToEntity>
       where L: Linked<FromEntity = Self::Entity> { ... }
    fn delete<'a, 'async_trait, A, C>(
        self,
        db: &'a C,
    ) -> Pin<Box<dyn Future<Output = Result<DeleteResult, DbErr>> + Send + 'async_trait>>
       where Self: IntoActiveModel<A> + 'async_trait,
             C: ConnectionTrait + 'async_trait,
             A: ActiveModelTrait<Entity = Self::Entity> + ActiveModelBehavior + Send + 'a + 'async_trait,
             'a: 'async_trait { ... }
    fn get_primary_key_value(&self) -> ValueTuple { ... }
}

The interface for Model, implemented by data structs

Required Associated Types

type Entity: EntityTrait

Required Methods

fn get(&self, c: <Self::Entity as EntityTrait>::Column) -> Value

Get the Value of a column from a Model

fn get_value_type(c: <Self::Entity as EntityTrait>::Column) -> ArrayType

Get the Value Type of a column from the Model

fn try_set( &mut self, c: <Self::Entity as EntityTrait>::Column, v: Value, ) -> Result<(), DbErr>

Set the Value of a Model field, return error if failed

Provided Methods

fn set(&mut self, c: <Self::Entity as EntityTrait>::Column, v: Value)

Set the Value of a Model field, panic if failed

fn find_related<R>(&self, _: R) -> Select<R>
where R: EntityTrait, Self::Entity: Related<R>,

Find related Models belonging to self

fn find_linked<L>(&self, l: L) -> Select<L::ToEntity>
where L: Linked<FromEntity = Self::Entity>,

Find linked Models belonging to self

fn delete<'a, 'async_trait, A, C>( self, db: &'a C, ) -> Pin<Box<dyn Future<Output = Result<DeleteResult, DbErr>> + Send + 'async_trait>>
where Self: IntoActiveModel<A> + 'async_trait, C: ConnectionTrait + 'async_trait, A: ActiveModelTrait<Entity = Self::Entity> + ActiveModelBehavior + Send + 'a + 'async_trait, 'a: 'async_trait,

Delete a model

fn get_primary_key_value(&self) -> ValueTuple

Get the primary key value of the Model

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.