sea_orm

Module query

Types and methods to perform queries

Re-exports

pub use crate::ConnectionTrait;

pub use crate::CursorTrait;

pub use crate::InsertResult;

pub use crate::PaginatorTrait;

pub use crate::SelectExt;

pub use crate::Statement;

pub use crate::StreamTrait;

pub use crate::TransactionTrait;

pub use crate::UpdateResult;

Macros

debug_query

Helper to get a raw SQL string from an object that impl QueryTrait.

debug_query_stmt

Helper to get a Statement from an object that impl QueryTrait.

Structs

Condition

Represents the value of an Condition::any or Condition::all: a set of disjunctive or conjunctive conditions.

DebugQuery

This structure provides debug capabilities

Delete

Defines the structure for a delete operation

DeleteMany

Perform a delete operation on multiple models

DeleteOne

A request to delete an ActiveModel.

DynIden

A prepared (quoted) identifier string.

Insert

Performs INSERT operations on a ActiveModel

InsertMany

Performs INSERT operations on many ActiveModels

Select

Defines a structure to perform select operations

SelectA

Implements the traits Iden for select alias

SelectB

Implements the traits Iden for select alias

SelectC

Implements the traits Iden for select alias

SelectFive

Perform a SELECT operation on 5 Models

SelectFour

Perform a SELECT operation on 4 Models

SelectSix

Perform a SELECT operation on 6 Models

SelectThree

Perform a SELECT operation on three Models

SelectThreeMany

Perform a SELECT operation on three Models with results consolidated

SelectTwo

Defines a structure to perform a SELECT operation on two Models

SelectTwoMany

Defines a structure to perform a SELECT operation on many Models

TopologyChain

A chain topology

TopologyStar

A star topology

TryInsert

Wrapper of Insert / InsertMany, treats “no row inserted/id returned” as a normal outcome.

Update

Defines a structure to perform UPDATE query operations on a ActiveModel

UpdateMany

Defines an UPDATE operation on multiple ActiveModels

UpdateOne

A request to update an ActiveModel.

ValidatedDeleteOne

A validated DeleteOne request, where the primary key is set and it’s possible to generate the right SQL condition.

ValidatedUpdateOne

A validated UpdateOne request, where the primary key is set and it’s possible to generate the right SQL condition.

Values

Enums

JoinType

Join types

JsonValue

Represents any valid JSON value.

Order

Ordering options

Value

Value variants

Traits

ColumnAsExpr

Extending IntoSimpleExpr to support casting ActiveEnum as TEXT in select expression

ConditionalStatement

EntityOrSelect

Entity, or a Select; to be used as parameters in LoaderTrait

ExprTrait

“Operator” methods for building expressions.

IntoSimpleExpr

Performs a conversion to SimpleExpr

LoaderTrait

This trait implements the Data Loader API

OrderedStatement

QueryFilter

Perform a FILTER opertation on a statement

QueryOrder

Performs ORDER BY operations

QuerySelect

Abstract API for performing queries

QueryTrait

A Trait for any type performing queries on a Model or ActiveModel

Topology

Topology of multi-joins