sea_orm
Module query

Holds types and methods to perform queries

Re-exports

pub use crate::ConnectionTrait;

pub use crate::CursorTrait;

pub use crate::InsertResult;

pub use crate::PaginatorTrait;

pub use crate::Statement;

pub use crate::StreamTrait;

pub use crate::TransactionTrait;

pub use crate::UpdateResult;

Macros

debug_query

Helper to get a raw SQL string from an object that impl QueryTrait.

debug_query_stmt

Helper to get a Statement from an object that impl QueryTrait.

Structs

Condition

Represents the value of an Condition::any or Condition::all: a set of disjunctive or conjunctive conditions.

DebugQuery

This structure provides debug capabilities

Delete

Defines the structure for a delete operation

DeleteMany

Perform a delete operation on multiple models

DeleteOne

Perform a delete operation on a model

Insert

Performs INSERT operations on a ActiveModel

Select

Defines a structure to perform select operations

SelectA

Implements the traits Iden and IdenStatic for a type

SelectB

Implements the traits Iden and IdenStatic for a type

SelectTwo

Defines a structure to perform a SELECT operation on two Models

SelectTwoMany

Defines a structure to perform a SELECT operation on many Models

TryInsert

Performs INSERT operations on a ActiveModel, will do nothing if input is empty.

Update

Defines a structure to perform UPDATE query operations on a ActiveModel

UpdateMany

Defines an UPDATE operation on multiple ActiveModels

UpdateOne

Defines an UPDATE operation on one ActiveModel

Values

Enums

JoinType

Join types

JsonValue

Represents any valid JSON value.

Order

Ordering options

Value

Value variants

Traits

ConditionalStatement

EntityOrSelect

Entity, or a Select; to be used as parameters in LoaderTrait

IntoSimpleExpr

Performs a conversion to SimpleExpr

LoaderTrait

This trait implements the Data Loader API

OrderedStatement

QueryFilter

Perform a FILTER opertation on a statement

QueryOrder

Performs ORDER BY operations

QuerySelect

Abstract API for performing queries

QueryTrait

A Trait for any type performing queries on a Model or ActiveModel

SelectColumns

Select specific column for partial model queries

Type Aliases

DynIden