sea_orm

Struct DatabaseConnection

#[non_exhaustive]
pub struct DatabaseConnection {
    pub inner: DatabaseConnectionType,
}

Handle a database connection depending on the backend enabled by the feature flags. This creates a connection pool internally (for SQLx connections), and so is cheap to clone.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

inner: DatabaseConnectionType

DatabaseConnection used to be a enum. Now it’s moved into inner, because we have to attach other contexts.

Implementations

impl DatabaseConnection

pub fn as_mock_connection(&self) -> &MockDatabaseConnection

Generate a database connection for testing the Mock database

Panics

Panics if DbConn is not a mock connection.

pub fn into_transaction_log(self) -> Vec<Transaction>

Get the transaction log as a collection Vec<crate::Transaction>

Panics

Panics if the mocker mutex is being held by another thread.

impl DatabaseConnection

pub fn as_proxy_connection(&self) -> &ProxyDatabaseConnection

Generate a database connection for testing the Proxy database

Panics

Panics if DbConn is not a proxy connection.

impl DatabaseConnection

pub fn get_database_backend(&self) -> DbBackend

Get the database backend for this connection

Panics

Panics if DatabaseConnection is Disconnected.

pub fn get_schema_builder(&self) -> SchemaBuilder

Creates a SchemaBuilder for this backend

pub fn set_metric_callback<F>(&mut self, _callback: F)
where F: Fn(&Info<'_>) + Send + Sync + 'static,

Sets a callback to metric this connection

pub async fn ping(&self) -> Result<(), DbErr>

Checks if a connection to the database is still valid.

pub async fn close(self) -> Result<(), DbErr>

Explicitly close the database connection. See Self::close_by_ref for usage with references.

pub async fn close_by_ref(&self) -> Result<(), DbErr>

Explicitly close the database connection

impl DatabaseConnection

pub fn get_mysql_connection_pool(&self) -> &MySqlPool

Get sqlx::MySqlPool

Panics

Panics if DbConn is not a MySQL connection.

pub fn get_postgres_connection_pool(&self) -> &PgPool

Get sqlx::PgPool

Panics

Panics if DbConn is not a Postgres connection.

pub fn get_sqlite_connection_pool(&self) -> &SqlitePool

Get sqlx::SqlitePool

Panics

Panics if DbConn is not a SQLite connection.

pub fn get_mssql_connection_pool(&self) -> &MsSqlPool

Get sqlz::MsSqlPool

Panics

Panics if DbConn is not a MSSQL connection.