sea_orm
Trait TryGetable

pub trait TryGetable: Sized {
    // Required method
    fn try_get_by<I: ColIdx>(
        res: &QueryResult,
        index: I,
    ) -> Result<Self, TryGetError>;

    // Provided methods
    fn try_get(
        res: &QueryResult,
        pre: &str,
        col: &str,
    ) -> Result<Self, TryGetError> { ... }
    fn try_get_by_index(
        res: &QueryResult,
        index: usize,
    ) -> Result<Self, TryGetError> { ... }
}

An interface to get a value from the query result

Required Methods

fn try_get_by<I: ColIdx>( res: &QueryResult, index: I, ) -> Result<Self, TryGetError>

Get a value from the query result with an ColIdx

Provided Methods

fn try_get(res: &QueryResult, pre: &str, col: &str) -> Result<Self, TryGetError>

Get a value from the query result with prefixed column name

fn try_get_by_index( res: &QueryResult, index: usize, ) -> Result<Self, TryGetError>

Get a value from the query result based on the order in the select expressions

Object Safety

This trait is not object safe.